Move watcher tracking strategies to separate classes
so they are easier to test.
