+++ subprojects/snapshots/src/main/java/org/gradle/internal/vfs/watch/FileWatcherRegistry.java
@@ -16,17 +16,12 @@
 
 package org.gradle.internal.vfs.watch;
 
-import org.gradle.internal.snapshot.CompleteFileSystemLocationSnapshot;
-import org.gradle.internal.snapshot.SnapshotHierarchy;
-
 import java.io.Closeable;
-import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.Collection;
 import java.util.Optional;
 
-public interface FileWatcherRegistry extends Closeable, SnapshotHierarchy.SnapshotDiffListener {
+public interface FileWatcherRegistry extends Closeable {
 
     interface ChangeHandler {
         void handleChange(Type type, Path path);
@@ -41,20 +36,7 @@ public interface FileWatcherRegistry extends Closeable, SnapshotHierarchy.Snapsh
         INVALIDATED
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @throws WatchingNotSupportedException when the native watchers can't be updated.
-     */
-    @Override
-    void changed(Collection<CompleteFileSystemLocationSnapshot> removedSnapshots, Collection<CompleteFileSystemLocationSnapshot> addedSnapshots);
-
-    /**
-     * Changes the must watch directories, e.g. when the same daemon is used on a different project.
-     *
-     * @throws WatchingNotSupportedException when the native watchers can't be updated.
-     */
-    void updateMustWatchDirectories(Collection<File> updatedWatchDirectories);
+    FileWatcherUpdater getFileWatcherUpdater();
 
     /**
      * Get statistics about the received changes.
